New and updated functions getCurrentRepo, updateRepoAuto updateRepo.
Function getCurrentRepo returns name of the current working dir if
it is in the list of CLONE and MAINSTREAM (either one of them).
Used by the function updateRepo.

Function updateRepoAuto pulls updates for the repository, provided
that it is called from within the repository working directory.
Updates will be pulled for the current branch, if no argument is
passed to the function.

Function updateRepo is a human double-checker before calling
function updateRepoAuto.